rag_eval_toolkit/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-eval-toolkit"
version = "0.1.0"
description = "A complete RAG experimentation framework for evaluating and optimizing RAG architectures"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Dunderscore AI" }
]
keywords = ["rag", "evaluation", "llm", "retrieval", "nlp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.0",
    "ragas>=0.3.5",
    "datasets>=2.14.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
graphrag = [
    "lightrag-hku>=0.1.0",
]

[project.scripts]
rag-eval = "rag_eval.cli:main"

[project.urls]
Homepage = "https://github.com/dunderscore-ai/rag-eval-toolkit"
Documentation = "https://github.com/dunderscore-ai/rag-eval-toolkit#readme"
Repository = "https://github.com/dunderscore-ai/rag-eval-toolkit"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_eval"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
